Send a more believable value for the syndication base date
The W3C feed validator[1] takes issue with a 1901 start date for
syndication; certainly, being before the UNIX epoch may cause problems
on some RSS implementations.

Switch to using "the last midnight" as the base time.  As the provided
frequency is "every hour", any timestamp which is on-the-hour should be
sufficient.  While midnight does not historically exist for _every_ day,
it has for all recent days in the current calendar system -- and, more
importantly, most probably will exist for all future days in the current
calendar system.  Unlike "truncate to the last hour," it is also unique
even at DST boundaries.

[1] http://validator.w3.org/feed/