fix error titlebox styles for rudder theme
This fixes multiple styling mistakes for the rudder theme:

The "Login failed" message is displayed in a yellow box with red
border instead of a red box without a border.

The horrible "Article needs setup" message and search results parsing
error message. I think the "error-titlebox" class was a mistake because
it was only defined in rudder. Use the "error" class here and wrap it in
a div.results as in other places. As the "div.results .titlebox-content"
is optimized for display lists as a result, wrap the content of the above
messages in a ul (this is for the other themes beside rudder).