Explicitly encode the HTML before passing it to external programs
HTML::FormatExternal prints the data that it recieves to the external
program, without explicitly applying an encoding -- that is, it takes
bytes, not characters.  When perl attempts to use "print" to send these
to the external program, it (correctly) produces "wide character in
print" warnings.

Explicitly encode the HTML to format as UTF-8 before sending it over the
wire; this makes perl's internal representation of the characters
irrelevant, and squashes the wide character warnings.