Be explicit about encodings to and from the external formatters
HTML::FormatExternal supports input_charset and output_charset
parameters controlling how it the input is encoded, and the encoding
that the formatter should attempt to output; otherwise, the formatter
attempts to determine based on the <meta> tag.  Many of the snippets
passed to ConvertHTMLToText do not have meta tags -- and, more
importantly, in all cases they have already been transcoded to UTF-8,
making the meta tag possibly incorrect.  Fortunately, the input_charset
argument is documented to override the charset found in the <meta> tag,
if any.

Not all formatters have the same support, of course.  The formatters RT
allows, and their encoding support, are (in preference order):
 * w3m supports arbitrary input and output character sets.
 * elinks supports arbitrary input and output character sets.
 * html2text supports only latin-1 input, and has no control over output
   character set.
 * links supports arbitrary input and output character sets.
 * lynx supports specifying a default input character set, but this does
   not override a <meta> tag.  It supports arbitrary output character
   sets.

Installs with only html2text or lynks are thus likely to suffer from
encoding problems; move them to be lowest on the priority list.  They
are still listed because mis-encoded mail is still superior to the
possibly-empty mail that the "core" formatter would send.

Tests need adjustment because EmailOutputEncoding is a suggestion -- if
the content contains characters which the given character set cannot
express, the character set is left as the internal default, UTF-8.  In
the case of w3m, an <hr /> tag is rendered as a series of code point
0x2501 ("BOX DRAWINGS HEAVY HORIZONTAL"), which is not in ISO-8859-1; as
such, the text/plain part is sent in UTF-8, not ISO-8859-1.