Only search Content, not Content and Subject, for better indexing
A TicketSQL query of « Subject LIKE 'word' OR Content LIKE 'word' »
generates the following SQL (simplified):

    SELECT DISTINCT main.*
      FROM Tickets main
           JOIN Transactions Transactions_1
             ON ( Transactions_1.ObjectType = 'RT::Ticket' )
            AND ( Transactions_1.ObjectId = main.id )
      LEFT JOIN Attachments Attachments_2
             ON ( Attachments_2.TransactionId = Transactions_1.id )
     WHERE main.Subject LIKE '%word%'
        OR Attachments_2.Content MATCHES 'word'

This leaves both MySQL and Postgres unable to use their full-text-search
indexes on the Content search, as their query planner cannot separate
this unto a union of two queries which are both well-indexed.  Instead,
both databases elect to perform sequential scans of all three involved
tables.  While the full-text index does speed this process, the
sequential scans nonethless cause the query to take noticable time.

For the common case of simple search with FTS enabled, default to
searching the Content rather than suffer the performance penalties of
searching both.  In nearly all cases, the words of the subject are
contained within the words of the content; later work will ensure that
the full-text-search index contains the information from the subject,
allowing a performant but fully inclusive search.