allow to search the string from anywhere in the word
by default, Chosen's search matches starting at the beginning of a word.
setting this option to true allows matches starting from anywhere within a word.
This is especially useful for options that include a lot of special characters
or phrases in ()s and []s.

I believe it's good to be enabled by default