Only store text as quoted-printable when the bytes not invalid UTF-8
The existing check for Encode::is_utf8() checks if the UTF8 flag is on
_and_ that the internal representation is consistent.  Since the UTF8
flag is almost certainly off, as _EncodeLOB takes bytes, not characters,
this causes the quoted-printable encoding to be chosen even for byte
sequences which are valid UTF-8.

Switch to checking if the byte sequence is valid UTF-8; if so, upgrade
it to characters and store it verbatim.  If it is not, then store it as
quoted-printable as before.  This increases the set of content which is
stored with encoding "none", which eases direct comprehension of the
data.