allow HTML5 <s> and <del> tags
The <s> and <del> tags replaced in HTML5 the old <strike> tag.

The <s> tag specifies text that is no longer correct, accurate or
relevant.
The <del> tag defines text that has been deleted from a document.

We still need to support the <strike> tag as the current used CKEditor
uses it. Newer versions of CKEditor use the <s> tag.