Add a new 'BodyClass' parameter to Elements/Header
Add a new 'BodyClass' parameter to Elements/Header to let extensions
(such as RTIR) pass in an explicit body class to all pages they
handle with a callback, so they can more easily style only their
own pages.

Extensions can add their own strings to this parameter by pushing
elements onto it, via the "Head" callback

We use the parameter as $ARGS{'BodyClass'} rather than @BodyClass
because the @BodyClass version is instantiated before the callback
gets a chance to add values to it.