Drop the severity of known indexing failures to "info"
These failures are known failure mores that cannot easily be addressed.
While the failure merits logging, as they are not unexpected in most
installs, there is no direct solution and they can, in most cases, be
ignored.  As such, "info" is a fine log level for a not-unexpected and
unhandle-able error.