Avoid treating attached GPG keys named *.asc as encrypted attachments
RT::Crypt::GPG looks for RFC3156 parts, as well as files with matching
.sig parts.  Any remaining attachments named .pgp or .asc are assumed to
be attached, encrypted files.

However, there is one common case which this is incorrect for --
attached public key files, which GPG exports with .asc, by default.
When RT attempts to "decrypt" these files, GPG responds:

    gpg: decrypt_message failed: Unexpected error

..or, for gpg2:

    gpg: decrypt_message failed: unexpected data

This results in the email sender receiving a "Error: bad encrypted data"
email, which is especially puzzling if their mail did not contain any
encrypted parts.

Skip .asc/.pgp parts which have content-type application/pgp-keys.
While this will not reduce the false-positives entirely (for mail
clients that do not set the content-type correctly), it is prefereable
to have more false-positives than false negatives.  Better testing is
impossible without examining the contents of the MIME part, which may be
quite memory-intensive.