Generalize Owner logic in Shredder to any Single role group
Extensions, like Assets, may add other relations which are Single role
groups -- that is, they are expected to always have one member in them,
even if that value is the Nobody user.

Switch from hard-coded checks on Owner to working on all Single role
groups; this allows users to be shredded in systems with Assets, and the
GroupMembers and CachedGroupMembers tables to remain consistent.