Only use CreateTicket line as a CF name if not also a core field
Every field CreateTickets loads is treated as potential CF; so, in the
presence of a "Status" CF, the CF was set, in addition to the core
Status field.

Once a field from %args has been used to fill %ticketargs, remove it,
such that it is not viewed as a potential custom field name.  To
accomplish this, handling of "deferred" items (including links and
Status) are moved before CF handling, so they can remove values they
consume.

Status is additionally special; it is both deferred _and_ set in
%ticketargs.  As such, it is temporarily put back into %args before
GetDeferred is called.

Fixes: I#18145