Widen IP custom fields to be wide enough for full IPv6 addresses
IP address custom fields get the default width for custom field
editing.  While this is wide enough for IPv4 addresses, it is not
sufficiently wide enough for full IPv6 addresses, which are 45
characters long.

Use CSS to widen the text input for all IP custom field entries. The
size of 24em was chosen based on testing in Safari, Firefox, Chrome and
IE.

Fixes: I#24565