Increase the bcrypt cost by a factor of two
The bcrypt key derivation function contains a tuning parameter, the
number of "rounds" to run, which allows the same algorithm to have
increased computational cost, to prevent brute force attacks, as
computers grow faster.

Moore's Law estimates a doubling in the number of transistors every 18
months, which is, in broad strokes, also similar to a doubling in
computational power every 18 months.  RT introduced the use of bcrypt in
September 2013; as such, the complexity of key derivation is now due to
be doubled.  This is done by increasing the number of rounds by one.

Old passwords (with lower costs) will still work, but will be upgraded
as soon as a user logs in.