Flag mentions of "attachment" when there are none
This is an imperfect (but hopefully good-enough) solution to the common
problem of forgetting to add attachments to tickets. There are lots of
corner cases handled (like warning when you add an attachment but later
check its ticky box to delete it) but there are a few that are less
tractable (like when you quote a transaction that mentioned the word
"attachment"—which disables this feature—and then forget to add your
own attachment in reply)

This feature does not forbid you from creating tickets or anything like
that. It merely provides a hint message in the UI.

Fixes: I#30951