consistent "now" value on ticket create
Without this, creating a ticket with a open status, Started is set to
now in RT::Ticket->Create and Created is set to now in
RT::Record->Create, where the later method call can be a second later
(actually on the next second) than the former one. This results in
Started is one second before Created.

Avoid this by setting Created on RT::Ticket->Create.

Also use a consistent now value for other date values, if they are set
to now.