show the queue id if the user can't see the queue name
Show the queue id if the user doesn't have sufficient rights to see
the queue name.

If you move a ticket from queue q1 to q2 and the user have SeeQueue (and
ShowTicket) on q2 but not on q1, the descriptions shows previously
'Queue changed from to q2', which isn't very helpfull. Showing the queue
id gives the user a hint to tell the admin which queue he can't see.

This makes the queue display consistent with /Ticket/Elements/ShowQueue.

This also fixes the following warnings:
Use of uninitialized value $_[2] in join or string
Use of uninitialized value $_[3] in join or string