Break out the time worked on a ticket per user
Under the total time worked in the ticket basics on the ticket display
page, display the time that each user has worked on the ticket. A scrip
is used to cache the time worked per user in a ticket attribute.

For new installs, default to the behavior enabled. For upgrades, install
the scrip disabled.