add highlighting functionality to Edit Rights admin pages
Use CSS to indicate when both category and principal tabs have rights
granted "underneath" them. This will visually help the administrator
know where unseen rights are granted in the admin UI. When all
checkboxes for a given category or principal have been unchecked (rights
revoked, thought still need to be committed), then the category and/or
principal tab will return to its unhighlighted state.

In order to keep the original text padding the same as how it is with
the highlighting, we need to tweak the padding of a couple of CSS
entries.