Avoid version 1.01 of DateTime::Locale
DateTime::Locale 1.01 included some changes to locale parsing
including a change in behavior such that en-us, which previously
loaded as a locale, no longer worked. This was updated in
version 1.02. See this ticket for details:

https://rt.cpan.org/Public/Bug/Display.html?id=110244

This was observed in version 1.01, but version 1.00 was the first
to include the large set of changes, so avoid that version as
well.

Fixes: I#31542