Update detection of find version in configure
Commit 4028723 added code to detect the version of find
because Debian is changing to a newer version of GNU
find with different syntax than the traditional find.
However, the change in that commit no longer worked on
Mac OS X, a BSD-type system.

Modify the find check to use the --version option supported
on GNU and use GNU syntax if found. Otherwise default to the
previous BSD syntax which was the previous setting for all
systems.