fix charts with QueueCF or TxnCF in query
RT::Report::Tickets (charts) don't have CustomFieldJoins registered.
We have to use the CustomFieldJoins from RT::Tickets (search)
here.

Without this, creating a chart with a QueueCF in the query fails with:
Couldn't parse query: We don't know how to join for LookupType RT::Queue

See also test from previous commit.
Same applies to charts with TxnCF in the query.

This was missed for TxnCF here 70f502c, and later for QueueCF here
7ef6480.