Allow sorting in dashboards for objects with a default order by
The logic in /Elements/CollectionList checks for a default
sort order in the object passed in unless you explicitly set
AllowSorting. By default, dashboards rendering a saved search
do not pass this. This is not be a problem for tickets
as they don't have default search orders. However,
other objects (like assets) default to sorting by Name,
so the sort order from the saved search is ignored.

Provide a way to give preference to a passed on OrderBy over
the default object order by.