force .load to POST instead of GET
AJAXifying the preview scrips with .load() via .serialize() does a GET
request with the Display.html form, and the query string can end up too
long for the server to accept it.

This commit adds a helper to serialize the form as an object instead of
query string, which tells .load to POST instead of GET. It looks like
theres no built in jQuery interface to serialize a form as an object so
the helper is needed.

Fixes issues#31874