Call __Value for CF Type to avoid permission check
d3928d7b refactored canonicalization for CFs and,
recognizing that a newly created CF object would not have
sufficient context to pass the rights check, called
_Value('Type') rather than Type to avoid it. However, _Value
implements the 'SeeCustomField' rights check and was
therefore causing canonicalization to fail, resulting in
incorrect datetime CF value setting for non-root users.

Call __Value (note the extra underscore) which is the
underlying method that does not require the rights check.

Fixes: I#31674