Change ticket timer from ticking seconds to measuring duration
There are several problems with incrementing a counter every second.  It
boils down to the fact that setInterval is certainly not a guarantee,
it's more of a best effort. In particular, on mobile devices, browsers
will not waste precious battery and CPU cycles to run timers for tabs in
the background. So when you come back to your ticket timer after minutes
or hours of work, the browser will not have been calling setInterval,
causing you to lose all that time worked.

Rather than incrementing a seconds counter every second, we instead
track when the last unpause happened, and take its difference from the
current time. If the user pauses and unpauses repeatedly, we also track
"committed seconds". Any time we need the duration (for display in the
UI or for submitting as time worked) we sum the two.  The result is that
when the browser resumes calling setInterval after being backgrounded
away for hours, that duration will be accurate, since it's directly
measuring time passed rather than relying on setInterval having been
called regularly. setInterval is now used _only_ for updating the
display, so it can be called as frequently or rarely as the browser
decides.

Fixes: I#31707