Tweak spaces around signature to provide useful default whitespace
Most signatures don't have trailing newlines, additional newlines are
necessary after the signature to give whitespace before showing the
quote.  Additionally, having a blank line (or two) before the "-- \n"
means there's an obvious place to click and start typing.  With
signature-before-quote, this now gives (all examples use "." for a blank
line):
    .
    .
     --
     Foo
     Bar
    .
     On Oct 18, you wrote:
     > ...

This provides a blank line to click on and begin to type, as well as a
blank line thereafter to separate the quote from the signature.

With the standard signature-after setting, this now gives:
     On Oct 18, you wrote:
     > ...
    .
    .
    .
     --
     Foo
     Bar

...providing three blank lines in the middle.  This allows the user to
click on the second one and start typing, but leave whitespace before
and after their response.  Finally, if there's nothing to quote,
regardless of the preference:
    .
    .
     --
     Foo
     Bar

..with a similar click-on-the-first-line use case as described above.
All of the same behavior is preserved when HTML editing is enabled.