Add additional output to give feedback while running
If run with no arguments, Init sets the log level to
warning, so the user won't see any of the info messages
that display what the script is doing. Add a start and
end print to send something to the terminal.

Also add a log line for the ticket update step.