Refactor, reword, and reorder CPAN checking
Start by suggesting that the user try installing the appropriate
system package that would provide CPAN; if this is possible, this is a
one-command method of making `make fixdeps` simply work.

Keep "use a different tool" as the next option, but provide explicit
instructions on how to do so for cpanm; the previous example simply
reused `perl -MCPAN -e"install %s"` as its installation tool, which is
unlikely to work, since the whole error arises from `require CPAN`
failing.

Remove the suggestion of "configure CPAN" because it is the
compile-time loading of the CPAN module which is failing; as such,
`perl -MCPAN -e shell` cannot work, since (as above) we have just
ascertained that their perl cannot load CPAN.pm

Leave the bottom two options as-is, although the likelihood of the
user being able to install CPAN by hand, or being willing to download
every module by hand, is exceedingly low.