Store dependencies in cpanfile
Previously, dependencies were stored in a complicated hash for unclear
reasons, and traversed in random order.  Additionally, the
specification of them included unwieldy boilerplate and here-docs.

Switch to storing the dependency list in a `cpanfile`, and load it by
stubbing the small number of functions to produce a useful
datastructure.  This switches the format to the more expressive '>='
and '!=' syntax used in META.yml[1] files, condensing the "AVOID"
list away.

Because at least one of the database "features" is always required,
installing directly via `cpanm --install-deps .` is not suggested or
documented at this time.  Using the `cpanfile` format, however, opens
the door to using `carton` and similar toolchains to manage
dependencies.

[1] https://metacpan.org/pod/CPAN::Meta::Spec#Version-Ranges