Move database and deployment defaults into rt-test-dependencies
rt-test-dependencies is rewritten with most, but not all, of its
arguments set to their configure-time values.  However, it requires
that users pass in the appropriate values for their database and
deployment method -- the two most key configuration settings.  Running
rt-test-dependencies directly (instead of via `make testdeps`) can
thus lead to a false belief that all dependencies are satisfied,
because deployment and database dependencies are not being checked.

Move the dependency and deployment defaulting into
rt-test-dependencies, from the Makefile.  In case rt-test-dependencies
is being called with an explicit database or deployment type, these
defaults are only set if no explicit setting is observed on the
command line.