Only generate the external link if the attachment is external
Though the documentation for the method notes that it may return undef
if the object is not in external storage, the code does not follow
this.  In an RT which is configured for external storage, attachments
may not be external yet if they do not meet the ShouldStoreExternally
criteria, or if the rt-externalize-attachments cron job has not yet
moved them.

In the case of Amazon, the one storage engine capable of direct
linking, assuming they are external causes the entire contents of the
attachment to be returned as the "URL," which is inlined into the
page with unfortunate results.

Only ask the engine for its external storage link if it is, in fact,
stored externally.