In tests, call PostLoadCheck to enable Externalauth
02b06597 made the ExternalAuth meta-setting automatically enabled via
PostLoadCheck, and 8946b1b8 removed the code in tests that set it.
Because of the bug fixed in 43c13180, however, the ExternalAuth flag
was always enabled, making tests pass.

Post-43c13180, the PostLoadCheck is never called after the ->Set calls
inside a testfile.  As such, the ExternalAuth meta-setting is never
enabled, and all external auth tests fail.

Explicitly call ->PostLoadCheck after each block of
ExternalAuth-related ->Set calls, to trigger the activation of the
ExternalAuth meta-setting.  Calling each configuration's PostLoadCheck
hook automatically after each ->Set would require careful adjustment
of the order of ->Set calls, which might be surprising behavior, as
ordering is not important in `RT_SiteConfig.pm`.