Make HTML::Gumbo required, and split scrubbing into its own class
HTML::Gumbo was previously an optional runtime dependency, which
allowed tables to not be scrubbed because it guaranteed that the
user-provided content couldn't "escape" the confines of the
transaction history.

Make HTML::Gumbo a required dependency, allowing tables in HTML for
all installations of RT.  In doing so, factor out the HTML-scrubbing
code into its own class, for compartmentalization.