Make Encode::HanExtra a required dependency
This makes RT better able to handle non-European mail.  While
previously an error was placed in the logs suggesting the installation
if a message was encounted which needed it, it was already too late
for that message.

There is functionally no cost to making it a required dependency; add
it.