Simplify file skipping
The movement of static resources into their own top-level directory
allows this logic to be simplified significantly -- especially when
combined with a fair number of "no longer relevant"s.

Changes must be made with an eye toward allowing this tool to be run
from within RT extensions, as well as RT itself (hence the inclusion
of any top-level `html` directory, for instance).

The following exceptions have been removed:

 - `lib/blib`: Nesting blib under lib is wrong, and doesn't happen
 - `lib/t/autogen`: Has never been relevant in core RT
 - `var`, `m4`, `local`: We skip these at the top level
 - `share/fonts`: Made unnecessary by limiting to `share/html`
 - `pot|po|gif|png|psd|jpe?g|svg|css|js`: Most of these exceptions
   came from static resouces or `share/po`, which are now skipped by
   limiting to `share/html`.  If any files matching these exist under
   `share/html`, they contain Mason code, so should be examined.
 - `,D` and `,B$`: Introduced in 5a89bf9d without comment; presumably
   editor cruft?
 - `extract-message-catalog`, `tweak-template-locstring`: Now live
   under `devel` so are exempt.