Refactor all extraction loops
This takes the common pattern of a while loop, optional dequoting of
values, and addition of the value to the list, and refactors them into
closures.  In doing so, it became apparent that the <&|/l&> form was
de-escaping when it should not have been; this has been fixed, though
it does not affect any existing locstrings.

There is no change to the extracted .pot file by this refactoring.