Move more error-checking into extraction-time
The aggregation done by doing so after all extraction is complete is
detrimental -- all occurrences of the errors will need to be fixed, so
only displaying one filename and line number forces the user to do
grepping.

The check for embedded newlines is moved to only trigger if the string
was double-quoted, averting a potential false-positive.