Allow tests to pass without a /usr/bin/sendmail
The installer tests include testing the page which configures outgoing
mail.  This page prompts for where the 'sendmail' binary is, and
checks that a file exists there.

The page defaults to `/usr/bin/sendmail`, which is installed on nearly
all systems, but may not on a system which only has the bar minimum
requirements to test RT -- on such systems, the test thus fails.

Provide a path to a known-existent binary which can suit.  All the
installer checks is that the file exists, simply provide the path to a
fake sendmail used elsewhere in tests.