Pass a fully-qualified path when requiring RT::Generated
As the comments a few lines down note, `$INC{'RT.pm'}` is not
guaranteed to be absolute.  By examining the path used by RT.pm, and
passing that to `require`, RT was attempting to skip the search of
@INC -- however, a relative path always searches @INC.

But the only means to find `lib/RT/Generated.pm` in @INC, in most
installs, is if the CWD is the root of an RT install, and @INC
contains `.`.  (Un)fortunately, perl's @INC nearly always was compiled
to contain `.` -- until CVE-2016-1238.

In response, many versions of perl now omit trailng `.` directories from
@INC, on a conditional or absolute basis.  Combined with the above,
this leads to failure to find RT::Generated.

Make the path passed to `require` always be absolute, to skip @INC
dependence entirely.