Stop removing space before 2nd and later MIME encoded-words
RFC 2047 is clear that only whitespace between encoded-word tokens is
to be stripped; space before them is significant.  However, RT trimmed
whitespace before all but the first encoded-word.  e541307d6e
introduced this trimming, along with a number of other suspect
internationalization fixes.

Remove this space trimming, document the absolutely correct `tr/_/ /`
replacement, and add a test for this behavior.