Hide scrip custom code fields when appropriate
The textareas for condition, action prepare, and action commit code
only make sense for scrips that use "User Defined" condition or action.
This commit shows or hides each textarea (using JavaScript) based on
whether the scrip's current configuration will use its contents. This
is meant to reduce user confusion about when these textareas take
effect, and significantly reduce the length of the page in the common
case of no "User Defined" code.

To reinforce to users the relationship between the dropdown and its
custom code textarea(s), we've moved each textarea up into the form,
right below to its dropdown. If the fields were to remain in two
separate sections, then users wouldn't notice the show/hide action at
a distance and would wonder why the textareas appear only some of the
time. This also reduces the perceived complexity of the scrip
create/modify page down to one consistent form, rather than two
confusingly-related sections, each with its own Save Changes button.

In order to ease the transition for custom deployments, a new method
(IsUserEditable) is added to RT::Condition and RT::Action to allow
users to easily inform RT of condition or action objects they have
made user editable.

Add Condition method to RT::ScripCondition for easily obtaining the
RT::Condition object, simillar to the Action method on RT::ScripAction.
Add lazy loading check to RT:ScripAction->Action.

Now that we've moved the textareas to be inline with the rest of the
form, there arises a new problem where specifying lots of code would
cause the rest of the page (e.g. Template selection) to scroll
offscreen. So adapting the size of the textbox to how many lines are
in the provided code (+3 lines for buffer) has been replaced with a
constant of 6 rows. Users will still be able to use their browser's
textarea resize tool to make the code entry fields longer.

There is no animation for the show/hide actions because jQuery's
slideUp and slideDown animations cannot handle table tags.

Fixes: I#32260