Disable inline edit for bulk update
Regardless of whether inline edit on bulk update even makes sense from a
UI perspective, the entire bulk update page is contained in a <form>
tag, and each inline edit widget uses its own <form> tag. This is
problematic because HTML doesn't support nested forms. While this could
be worked around by having inline edit reuse the form tag when possible
and providing it when not, the additional complexity is not worth the
(dubious) benefit.