Better support inline edit of datepicker fields
Focus leaves the inline edit form when interacting with the datepicker,
as that element lives as a direct child of the <body> tag. So when we're
using a datepicker, we can't use focusout as a hint that we should submit
the inline edit.

Instead we listen for the datepicker:close event for our hint that the user
is done picking a date.

Since the datepicker's DOM elements live outside of the inline edit <form>
tag, we don't receive select onchange events for selecting the month/year,
etc, so we don't have to filter out such events from our submit hint
heuristics.

Finally, because datepicker can't handle reused ids, we explicitly suppress
them from inline edit datepicker fields. Since these fields are new there's
less concern about backcompat.