Clean up transactions and records for semi-migrated Links
When serializing only a subset of records out of an RT instance,
special consideration is needed for any RT::Link record relating
some record in the serialized subset with some record not in the
serialized subset. This commit drops such RT::Link records, and the
related AddLink- and DeleteLink-type RT::Transaction records replace the
missing link with a "(not migrated)" note.

To inspect whether both sides of a link will be serialized, each record
class's Serialize method needs a way to query the RT::Migrate::Serializer
object orchestrating the serialization.

To enable an RT::Link record opting out of being serialized,
returning an empty hash from the Serialize method is now treated as a
special case to skip that record.