Add a --hyperlink-unmigrated option to rt-serializer
When serializing only a subset of records out of an RT instance, special
consideration is needed for any RT::Link record relating some record in
the serialized subset with some record not in the serialized subset. By
default, such RT::Link records are dropped.

With this commit, such RT::Link records will become hyperlinks instead,
and the AddLink- and DeleteLink-type RT::Transaction records will use
the hyperlink instead of the unmigrated local record.