Move onclick event to Create linked ticket anchor
On assets with a long history, it was possible to click
the Create linked ticket link in the Actions menu before
the page had fully loaded. In these cases, the jQuery
onclick event would not fire and the user would see the HTML
interstitial page rather than the modal for selecting the
queue to create the ticket.

Move the onclick event to the anchor tag itself to make sure
it displays the modal even if the page is not fully loaded.