Consider only query params provided by menu item for active status
This fixes a longstanding bug where, e.g. if you reply to a ticket, the
Display page menu item is no longer bolded. This is because the query
string doesn't match exactly, as in:

    expected: ?id=1
    got: ?id=1&results=hash

The new implementation considers only the query parameters that are
provided in the menu item, so any additional parameters (such as "results"
above) are not included in the comparison. The query parameters that are
provided in the menu item must be considered (as opposed to) because otherwise
you get lots of false positives, especially with the multitude of links
to /Ticket/Update.html with various different query parameters.

This also fixes the menu bolding to occur during pjax page loads, which add a
_pjax=1 query parameter, breaking the old implementation completely.