Improve password complexity configuration
This expands the basic MinimumPasswordLength configuration option into a
set of options under the new PasswordPolicy key. The new options allow
checks for a minimum number of various classes of character.

The web interface password fields are longer, to encourage use of a long
passphrase. Some basic guidance for selecting a good password is displayed.
The password requirements, as configured, are displayed to the user when
they are entering a new password.

Fixes: T#161950