Switch primary jQuery(function ...) registrations to handle pjax too
This is necessary because jQuery(function () { ... }) fires only once,
when the first page load is complete. When a pjax page load happens, the
jQuery(function () { ... }) handler is not invoked again because it's
still the same page load as far as the browser is concerned. So to run
JS for each pjax page load we must run them on pjax:success events as well.

According to https://api.jquery.com/ready/, jQuery(function () { ... }) is
equivalent to jQuery(document).ready(function () { ... }) (which is equivalent
to jQuery(document).on('ready', function () { ... })), so the timing of the
initial "page loaded" callbacks will remain unchanged.

This is required only for standalone .js files; HTML files that include a
    <script type="text/javascript">
        jQuery(function () { ... })
    </script>
will have such code executed even during pjax loads.

These changes unbreak the following JS-based features:
    - datepicker

    - ckeditor
    - user accordion (for "more about requestors")
    - chosen select box
    - combobox
    - attachment warning

    - autocomplete

    - disable queue checkboxes when apply globally is selected
    - linkify users in history
    - cascade-select custom fields
    - messagebox background color for dynamic change of reply/comment

    - assets accordion on ticket display
    - lightbox for asset create
    - lightbox for linked ticket create on assets

    - keep multiple input fields for the same custom field in sync

    - avoiding form double submission in Safari